pyproject.toml
[project]
name = "django-ninja-boilerplate"
version = "0.1.0"
description = ""
authors = [
    {name = "Bruno Nonogaki",email = "brunono@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django-ninja (>=1.5.0,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg (>=3.2.13,<4.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
ruff = "^0.12.11"
taskipy = "^1.14.1"
pytest-watch = "^4.2.0"
honcho = "^2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
extend-exclude = ['migrations', 'manage.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
services-up = "docker compose -f infra/compose-dev.yaml up -d"
services-stop = "docker compose -f infra/compose-dev.yaml stop"
services-down = "docker compose -f infra/compose-dev.yaml down"
create-env-dev = "ln -sf .env.development .env"
create-env-prod = "ln -sf .env.production .env"
run = 'task create-env-dev && task services-up && python infra/wait-for-postgres.py && python manage.py migrate && python manage.py runserver'
down = "pkill -f 'manage.py runserver'; docker compose -f infra/compose-dev.yaml down"
test = 'task create-env-dev && task services-up && python infra/wait-for-postgres.py && honcho start web test'
test-watch = 'pytest-watch'
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
pre_test = 'task lint'
post_test = 'coverage html'
